+++ include/jsoncons_ext/jmespath/jmespath.hpp
@@ -1829,7 +1829,7 @@ namespace jmespath {
                         {
                             return *resources.create_json(sval);
                         }
-                        jsoncons::detail::to_double_t to_double;
+                        jsoncons::detail::chars_to to_double;
                         try
                         {
                             auto s = arg0.as_string();